neural_db/requirements.txt
openai==0.27.6
tools==0.1.9
pymupdf==1.22.3
markupsafe==2.0.1
pandas==2.0.1
numpy==1.24.3
sentencepiece==0.1.99
protobuf==3.20
pillow==9.5.0
frontend==0.0.3
langchain==0.0.171
paper-qa==1.6.0
urllib3<2
nltk==3.8.1
python-docx
thirdai==0.7.9
streamlit
opencv-python-headless
easyocr
Pillow
